Move autoUpdater to modules folder
